chaning header size + little refactoring
