+++ src/components/Products (Product)/products.css
@@ -20,7 +20,7 @@
 }
 
 .product-header-txt {
-  font-family:'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
+  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
   font-size: 23px;
 }
